move loading of the cypress helper into command executor
