+++ .gitignore
@@ -2,3 +2,4 @@
 Gemfile.lock
 .ruby-version
 spec/examples.txt
+.idea
