+++ lib/cypress/command_executor.rb
@@ -1,7 +1,10 @@
+require 'cypress/configuration'
+
 module Cypress
   # Middleware to handle cypress commands and eval
   class CommandExecutor
     def self.load(file,command_options = nil)
+      load_cypress_helper
       file_data = File.read(file)
       eval file_data
     rescue => e
@@ -9,8 +12,18 @@ module Cypress
       raise e
     end
 
+    def self.load_cypress_helper
+      cypress_helper_file = "#{configuration.cypress_folder}/cypress_helper"
+      return unless File.exist?("#{cypress_helper_file}.rb")
+      Kernel.require cypress_helper_file
+    end
+
     def self.logger
-      Cypress.configuration.logger
+      configuration.logger
+    end
+
+    def self.configuration
+      Cypress.configuration
     end
   end
 end
