+++ spec/cypress/command_executor/cypress_helper.rb
@@ -1,0 +1,6 @@
+class DummyTest
+  # @return [Array]
+  def self.values
+    @values ||= []
+  end
+end
