+++ spec/cypress/command_executor/test_command.rb
@@ -1,0 +1,1 @@
+DummyTest.values << 'hello'
