+++ spec/cypress/command_executor/test_command_with_options.rb
@@ -1,0 +1,1 @@
+DummyTest.values << command_options
