+++ spec/cypress/command_executor_spec.rb
@@ -1,14 +1,8 @@
 require_relative '../../lib/cypress/command_executor'
 
-class DummyTest
-  # @return [Array]
-  def self.values
-    @values ||= []
-  end
-end
-
 RSpec.describe Cypress::CommandExecutor do
   describe '.load' do
+    let(:folder) { "#{__dir__}/command_executor" }
     subject { described_class }
 
     def executor_load(*values)
@@ -16,23 +10,24 @@ RSpec.describe Cypress::CommandExecutor do
     end
 
     before do
-      DummyTest.values.clear
+      Cypress.configuration.cypress_folder = folder
+      DummyTest.values.clear if defined?(DummyTest)
     end
 
     it 'runs test command' do
-      executor_load("#{__dir__}/test_command.rb")
-      expect(DummyTest.values).to eq(['hello'])
+      executor_load("#{folder}/test_command.rb")
+      expect(DummyTest.values).to eq(%w(hello))
     end
 
     it 'runs test command twice' do
-      executor_load("#{__dir__}/test_command.rb")
-      executor_load("#{__dir__}/test_command.rb")
-      expect(DummyTest.values).to eq(['hello', 'hello'])
+      executor_load("#{folder}/test_command.rb")
+      executor_load("#{folder}/test_command.rb")
+      expect(DummyTest.values).to eq(%w(hello hello))
     end
 
     it 'runs command with options' do
-      executor_load("#{__dir__}/test_command_with_options.rb", 'my_string')
-      expect(DummyTest.values).to eq(['my_string'])
+      executor_load("#{folder}/test_command_with_options.rb", 'my_string')
+      expect(DummyTest.values).to eq(%w(my_string))
     end
   end
 end
