--- spec/cypress/test_command.rb
@@ -1,1 +1,0 @@
-DummyTest.values << 'hello'
