--- spec/cypress/test_command_with_options.rb
@@ -1,1 +1,0 @@
-DummyTest.values << command_options
