+++ spec/integrations/rails_5_2/.gitignore
@@ -1,0 +1,10 @@
+.bundle
+spec/node_modules
+spec/cypress.json
+spec/package.json
+spec/yarn.lock
+spec/cypress/
+config/initializers/cypress_on_rails.rb
+vendor/bundle
+tmp/pids
+tmp/cache
